Added giant again and configured damage
The ONLY necessary code for this to function is the edited behavior code and PlayerDamage having the takedamage script. IF THIS BREAKS GUNS, RENAME TAKEDAMAGE ON PLAYERDAMAGE TO SOMETHING ELSE. I suspect the naming conventions being identical might be why unity gets confused. Otherwise this should make little impact outside of the dmg