Fixed Return to main Menu and Timer
Now when you switch to the main menu it is usable. (Issue was that the cursor was being deactivated so I just had to activate the cursor in the startfunction of the main menu script.
The timer now shows only when there is 30 seconds left from a 2minute timer. Returns you to the homeworld.